requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
altair==5.4.0
annotated-types==0.7.0
anthropic==0.25.8
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
binpacking==1.5.2
bitsandbytes==0.43.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.153
botocore==1.34.153
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpickle==3.0.0
cn2an==0.5.22
cohere==5.6.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.0
cupy-cuda12x==12.1.0
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.7
datasets==2.20.0
datasketch==1.6.4
debugpy==1.8.5
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.2
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.25.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
duckduckgo_search==5.3.1b1
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
et-xmlfile==1.1.0
evalscope==0.5.2
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.13
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.5
fastjsonschema==2.20.0
fastrlock==0.8.2
feedparser==6.0.11
ffmpy==0.4.0
filelock==3.15.4
flash-attn==2.6.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
func_timeout==4.3.5
future==1.0.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.5
google-ai-generativelanguage==0.4.0
google-api-core==2.19.0
google-auth==2.29.0
google-generativeai==0.3.2
googleapis-common-protos==1.63.0
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.3
griffe==1.1.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
imageio==2.35.1
immutabledict==4.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.2
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
kubernetes==30.1.0
lagent==0.2.3
langchain==0.2.12
langchain-anthropic==0.1.4
langchain-chroma==0.1.0
langchain-cohere==0.2.0
langchain-community==0.2.11
langchain-core==0.2.28
langchain-experimental==0.0.64
langchain-google-genai==0.0.9
langchain-huggingface==0.0.3
langchain-openai==0.1.20
langchain-postgres==0.0.9
langchain-text-splitters==0.2.2
langdetect==1.0.9
langgraph==0.0.25
langsmith==0.1.96
lark==1.1.9
lazy_loader==0.4
Levenshtein==0.25.1
llvmlite==0.43.0
ltp==4.2.14
ltp-core==0.1.4
ltp-extension==0.1.13
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmengine==0.10.4
mmengine-lite==0.10.4
mmh3==4.1.0
modelscope==1.17.1
monotonic==1.6
mpmath==1.3.0
ms-opencompass==0.0.7
ms-swift==2.3.1
ms-vlmeval==0.0.6
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.5.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.18.1
openai==1.38.0
OpenCC==1.1.9
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
oss2==2.18.6
outlines==0.0.46
overrides==7.7.0
packaging==24.1
pandas==2.2.1
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pgvector==0.2.5
phx-class-registry==4.1.0
pillow==10.3.0
pip==24.2
platformdirs==4.2.2
plotly==5.23.0
ply==3.11
portalocker==2.10.1
posthog==3.5.0
prettytable==3.11.0
proces==0.1.7
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
Pympler==1.1
pynvml==11.5.0
pyparsing==3.1.2
PyPika==0.48.9
pypinyin==0.52.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
rank-bm25==0.2.2
rapidfuzz==3.9.6
ray==2.32.0
referencing==0.35.1
regex==2024.5.10
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.19.0
rsa==4.9
ruff==0.6.2
s3transfer==0.10.2
sacrebleu==2.4.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==69.5.1
sgmllib3k==1.0.0
shellingham==1.5.4
shtab==1.7.1
simple-ddl-parser==1.6.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.30
sqlglot==23.3.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.1
sty==1.0.6
sympy==1.13.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
timeout-decorator==0.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
transformers-stream-generator==0.0.5
triton==3.0.0
trl==0.9.6
typer==0.12.3
types-python-dateutil==2.9.0.20240821
types-requests==2.32.0.20240712
typing_extensions==4.11.0
typing-inspect==0.9.0
tyro==0.8.8
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
validators==0.33.0
vllm==0.3.3
watchdog==4.0.2
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.43.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xformers==0.0.27.post2
XlsxWriter==3.2.0
xtuner==0.1.23
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
